refactor(utilities): make `sleepSync` its own file
