+++ packages/utilities/src/index.ts
@@ -4,12 +4,12 @@ export * from './lib/chunk';
 export * from './lib/classExtends';
 export * from './lib/codeBlock';
 export * from './lib/cutText';
-export { debounce, type DebounceSettings, type DebouncedFunc } from './lib/debounce/debounce';
+export * from './lib/debounce/debounce';
 export * from './lib/deepClone';
 export * from './lib/filterNullAndUndefined';
 export * from './lib/filterNullAndUndefinedAndEmpty';
 export * from './lib/filterNullAndUndefinedAndZero';
-export { getDeepObjectKeys, type GetDeepObjectKeysOptions } from './lib/getDeepObjectKeys';
+export * from './lib/getDeepObjectKeys';
 export * from './lib/hasAtLeastOneKeyInMap';
 export * from './lib/inlineCodeBlock';
 export * from './lib/isClass';
@@ -20,7 +20,7 @@ export * from './lib/isNullOrUndefinedOrZero';
 export * from './lib/isNumber';
 export * from './lib/isObject';
 export * from './lib/isPrimitive';
-export { isThenable, type Thenable } from './lib/isThenable';
+export * from './lib/isThenable';
 export * from './lib/lazy';
 export * from './lib/makeObject';
 export * from './lib/mergeDefault';
@@ -37,7 +37,8 @@ export * from './lib/regExpEsc';
 export * from './lib/retry';
 export * from './lib/retrySync';
 export * from './lib/roundNumber';
-export { AbortError, sleep, sleepSync, type SleepOptions } from './lib/sleep';
+export * from './lib/sleep';
+export * from './lib/sleepSync';
 export * from './lib/splitText';
 export { throttle, type ThrottleFn } from './lib/throttle';
 export { toTitleCase, type ToTitleCaseOptions } from './lib/toTitleCase';
