+++ packages/utilities/src/lib/sleep.ts
@@ -28,6 +28,7 @@ export class AbortError extends Error {
 
 /**
  * Sleeps for the specified number of milliseconds.
+ * For a synchronous variant, see [sleepSync](./sleepSync.d.ts).
  * @param ms The number of milliseconds to sleep.
  * @param value A value with which the promise is fulfilled.
  * @see {@link sleepSync} for a synchronous version.
@@ -50,19 +51,4 @@ export function sleep<T = undefined>(ms: number, value?: T, options?: SleepOptio
 			timer.unref();
 		}
 	});
-}
-
-/**
- * Sleeps for the specified number of milliseconds synchronously.
- * We should probably note that unlike {@link sleep} (which uses CPU tick times),
- * sleepSync uses wall clock times, so the precision is near-absolute by comparison.
- * That, and that synchronous means that nothing else in the thread will run for the length of the timer.
- * @param ms The number of milliseconds to sleep.
- * @param value A value to return.
- * @see {@link sleep} for an asynchronous version.
- */
-export function sleepSync<T = undefined>(ms: number, value?: T): T {
-	const end = Date.now() + ms;
-	while (Date.now() < end) continue;
-	return value!;
 }
