+++ packages/utilities/tests/sleep.test.ts
@@ -1,4 +1,4 @@
-import { AbortError, sleep, sleepSync } from '../src';
+import { AbortError, sleep } from '../src';
 import { expectError } from './util/macros/comparators';
 
 const DOMException: typeof globalThis.DOMException =
@@ -49,20 +49,5 @@ describe('sleep', () => {
 				cause: 'test'
 			})
 		);
-	});
-});
-
-describe('sleepSync', () => {
-	test('GIVEN a number of ms THEN return after that time', () => {
-		const start = Date.now();
-		sleepSync(50);
-		expect(Date.now() - start).oneOf([50, 51]);
-	});
-
-	test('GIVEN a number of ms and a value THEN return after that time with the value', () => {
-		const start = Date.now();
-		const value = sleepSync(50, 'test');
-		expect(Date.now() - start).oneOf([50, 51]);
-		expect<string>(value).toBe('test');
 	});
 });
